// src/gen-ts/src/models/BankFee.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni.
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime'
import type { CrupdateExpenseMoney } from './CrupdateExpenseMoney'
import {
  CrupdateExpenseMoneyFromJSON,
  CrupdateExpenseMoneyFromJSONTyped,
  CrupdateExpenseMoneyToJSON,
  CrupdateExpenseMoneyToJSONTyped,
} from './CrupdateExpenseMoney'

/**
 *
 * @export
 * @interface BankFee
 */
export interface BankFee {
  /**
   *
   * @type {string}
   * @memberof BankFee
   */
  id?: string
  /**
   *
   * @type {CrupdateExpenseMoney}
   * @memberof BankFee
   */
  expense?: CrupdateExpenseMoney
  /**
   *
   * @type {string}
   * @memberof BankFee
   */
  bank_name?: string
  /**
   *
   * @type {string}
   * @memberof BankFee
   */
  description?: string
}

/**
 * Check if a given object implements the BankFee interface.
 */
export function instanceOfBankFee(value: object): value is BankFee {
  return true
}

export function BankFeeFromJSON(json: any): BankFee {
  return BankFeeFromJSONTyped(json, false)
}

export function BankFeeFromJSONTyped(json: any, ignoreDiscriminator: boolean): BankFee {
  if (json == null) {
    return json
  }
  return {
    id: json['id'] == null ? undefined : json['id'],
    expense: json['expense'] == null ? undefined : CrupdateExpenseMoneyFromJSON(json['expense']),
    bank_name: json['bank_name'] == null ? undefined : json['bank_name'],
    description: json['description'] == null ? undefined : json['description'],
  }
}

export function BankFeeToJSON(json: any): BankFee {
  return BankFeeToJSONTyped(json, false)
}

export function BankFeeToJSONTyped(
  value?: BankFee | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value
  }

  return {
    id: value['id'],
    expense: CrupdateExpenseMoneyToJSON(value['expense']),
    bank_name: value['bank_name'],
    description: value['description'],
  }
}